backend/requirements.txt
asgiref==3.11.0
dataclasses==0.6
Django==3.1.12
django-cors-headers==3.11.0
djangorestframework==3.15.1
dnspython==2.8.0
pillow==12.0.0
PyJWT==2.10.1
pymongo==3.11.4
python-dotenv==1.2.1
pytz==2025.2
sqlparse==0.2.4
google-generativeai
gunicorn
whitenoise
